% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta.R
\name{rbetA}
\alias{rbetA}
\title{beta distribution sampler}
\usage{
rbetA(n, a, b)
}
\arguments{
\item{n}{the samples number (numeric)}

\item{a}{the first parameter of the distribution (numeric)}

\item{b}{the second parameter of the distribution (numeric)}
}
\value{
a random sample of size \code{n}
}
\description{
a random sample of size 1000 from Beta(3,2) distribution
}
\examples{
\dontrun{
rrrr<- Rbeta(n,a,b)
}
}
