% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rray.R
\name{RraY}
\alias{RraY}
\title{Rayleigh distribution sampler}
\usage{
RraY(m, sigma)
}
\arguments{
\item{m}{the samples number (numeric)}

\item{sigma}{the parameter of the distribution (numeric)}
}
\value{
a random sample of size \code{n}
}
\description{
generate samples from Rayleigh density distribution
}
\examples{
\dontrun{
rrrr<- Rray(m,sigma)
}
}
