% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{construct a table}
\usage{
table2(x, y)
}
\arguments{
\item{x}{vector 1 (numeric)}

\item{y}{vector 2 (numeric)}
}
\value{
a random sample of size \code{n}
}
\description{
a fast version of table function
}
\examples{
\dontrun{
a<-c(1,2,3,4)
b<-c(2,3,4,5)
table(a,b)
}
}
